Unifying KS prop I/O Standardizing sample input/output

+++ ks_hl_spectrum/clover_info.c
@@ -23,20 +23,7 @@
    write_appl_w_prop_info  Writes supplementary information to the info file */
 
 /* Include files */
-#include "../include/config.h"  /* Keep this first */
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include "../include/complex.h"
-#include "../include/su3.h"
-#include "../include/macros.h"
-#include "lattice.h"
-#include "../include/comdefs.h"	/* definitions and variables for communications */
-#include "../include/io_lat.h"
-#include "../include/io_wb.h"
-#include "../include/generic.h"
-#include "../include/generic_wilson.h"
-#include "../include/dirs.h"
+#include "ks_hl_spectrum_includes.h"
 /*---------------------------------------------------------------------------*/
 
 /* Fill in the spin table of contents for the propagator header -
@@ -68,7 +55,7 @@ void build_w_prop_hdr(w_prop_header *wph)
 void write_appl_w_prop_info(FILE *fp)
 {
   char sums[20];
-  float gauge_fix_tol = GAUGE_FIX_TOL;
+  double gauge_fix_tol = GAUGE_FIX_TOL;
 
   /* Note that the file has already been opened and
      the required magic number, time stamp, and lattice
