+++ ks_hl_spectrum/twopt.c
@@ -1,4 +1,4 @@
-#include "ks_dyn_includes.h"
+#include "ks_hl_spectrum_includes.h"
 
 #define GAMMAFIVE -1
 #define eve  0
@@ -119,7 +119,7 @@ void KS_2pt_func(field_offset snk, field_offset src, int *g_snk, int n_snk,
   int i, t, my_x,my_y,my_z;
   site *s;
   complex trace, trace1, epx,epx1;
-  float pi, mom[3];
+  double pi, mom[3];
 
   su3_matrix *antiquark;
   wilson_propagator  *quark;
@@ -129,9 +129,9 @@ void KS_2pt_func(field_offset snk, field_offset src, int *g_snk, int n_snk,
      for(i=0;i<nt;i++) printf("%.9e %.9e\n", prop[i].real,prop[i].imag);*/
 
   pi = 4.0 * atan( 1.);
-  mom[0] = -2.*pi/(float)nx;  
-  mom[1] = -2.*pi/(float)ny;  
-  mom[2] = -2.*pi/(float)nz; 
+  mom[0] = -2.*pi/(double)nx;  
+  mom[1] = -2.*pi/(double)ny;  
+  mom[2] = -2.*pi/(double)nz; 
  
   FORALLSITES(i,s){
     t = s->t;
@@ -140,8 +140,8 @@ void KS_2pt_func(field_offset snk, field_offset src, int *g_snk, int n_snk,
     my_z = s->z;
   
     epx1.real = 0.0;
-    epx1.imag = (mom[0]*(float)p[0]*(float)my_x + mom[1]*(float)p[1]*(float)my_y 
-		+ mom[2]*(float)p[2]*(float)my_z);
+    epx1.imag = (mom[0]*(double)p[0]*(double)my_x + mom[1]*(double)p[1]*(double)my_y 
+		+ mom[2]*(double)p[2]*(double)my_z);
     epx = cexp(&epx1);
 
     antiquark = (su3_matrix *)F_PT(s, snk);
@@ -186,7 +186,7 @@ void All_KS_hl_prop(field_offset snk, field_offset src, complex **propagator)
   int i, j;
   int g_snk[4],g_src[4];
   int n_snk,n_src;
-  float temp;
+  double temp;
 
   int p000[3] ={0,0,0}; 
 
@@ -852,16 +852,16 @@ void All_KS_hl_prop(field_offset snk, field_offset src, complex **propagator)
 
 int calculate_stag_prop() /* return the C.G. iteration number */
 {
-  float mass_x2;
+  double mass_x2;
   register complex cc,cc2;
-  float finalrsq, th;
+  double finalrsq, th;
   register int i,x,y,z,t,icol,cgn;
   site *st;
 
 
   /* Fix ZUP Coulomb gauge - gauge links only*/
   rephase( OFF );
-  gaugefix(ZUP,(float)1.8,500,(float)GAUGE_FIX_TOL,
+  gaugefix(ZUP,(double)1.8,500,(double)GAUGE_FIX_TOL,
 	   F_OFFSET(tempmat1),F_OFFSET(tempvec[0]),0,NULL,NULL,0,NULL,NULL);
   rephase( ON );
